more gateway simplification
